+++ packages/html-reporter/src/index.tsx
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import type { HTMLReport, Stats } from './types';
+import type { HTMLReport } from './types';
 import type zip from '@zip.js/zip.js';
 // @ts-ignore
 import * as zipImport from '@zip.js/zip.js/lib/zip-no-worker-inflate.js';
@@ -23,6 +23,7 @@ import * as ReactDOM from 'react-dom';
 import './colors.css';
 import type { LoadedReport } from './loadedReport';
 import { ReportView } from './reportView';
+import { mergeReports } from './mergeReports';
 // @ts-ignore
 const zipjs = zipImport as typeof zip;
 
@@ -33,10 +34,10 @@ const ReportLoader: React.FC = () => {
       return;
     const shardTotal = window.playwrightShardTotal;
     const zipReport = new ZipReport();
-    const loadPromis = shardTotal ?
+    const loadPromise = shardTotal ?
       zipReport.loadFromShards(shardTotal) :
       zipReport.loadFromBase64(window.playwrightReportBase64!);
-    loadPromis.then(() => setReport(zipReport));
+    loadPromise.then(() => setReport(zipReport));
   }, [report]);
   return <ReportView report={report}></ReportView>;
 };
@@ -57,7 +58,7 @@ class ZipReport implements LoadedReport {
   async loadFromShards(shardTotal: number) {
     const readers = [];
     for (let i = 0; i < shardTotal; i++) {
-      const fileName = `/report-${i + 1}-of-${shardTotal}.zip`;
+      const fileName = `report-${i + 1}-of-${shardTotal}.zip`;
       const zipReader = new zipjs.ZipReader(new zipjs.HttpReader(fileName), { useWebWorkers: false }) as zip.ZipReader;
       readers.push(this._readReportAndTestEntries(zipReader));
     }
@@ -82,34 +83,3 @@ class ZipReport implements LoadedReport {
   }
 }
 
-function mergeReports(reports: HTMLReport[]): HTMLReport {
-  const [report, ...rest] = reports;
-
-  for (const currentReport of rest) {
-    currentReport.files.forEach(file => {
-      const existingGroup = report.files.find(({ fileId }) => fileId === file.fileId);
-
-      if (existingGroup) {
-        existingGroup.tests.push(...file.tests);
-        mergeStats(existingGroup.stats, file.stats);
-      } else {
-        report.files.push(file);
-      }
-    });
-
-    mergeStats(report.stats, currentReport.stats);
-    report.metadata.duration += currentReport.metadata.duration;
-  }
-
-  return report;
-}
-
-function mergeStats(toStats: Stats, fromStats: Stats) {
-  toStats.total += fromStats.total;
-  toStats.expected += fromStats.expected;
-  toStats.unexpected += fromStats.unexpected;
-  toStats.flaky += fromStats.flaky;
-  toStats.skipped += fromStats.skipped;
-  toStats.duration += fromStats.duration;
-  toStats.ok = toStats.ok && fromStats.ok;
-}
